change rate value type from int to decimal

+++ passenger-service/src/main/java/io/voitovich/yura/passengerservice/entity/Rating.java
@@ -3,6 +3,7 @@ package io.voitovich.yura.passengerservice.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -19,7 +20,7 @@ public class Rating {
     private UUID id;
 
     @Column(name="rate_value", nullable = false)
-    private int rateValue;
+    private BigDecimal rateValue;
 
     @Column(name="driver_id", nullable = false)
     private UUID driverId;
