+++ passenger-service/src/main/resources/db/changelog/v-1.0/changesets/sql-files/create-rating-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS rating (
       id UUID PRIMARY KEY,
-      rate_value INT NOT NULL,
+      rate_value DECIMAL NOT NULL,
       driver_id UUID NOT NULL,
       passenger_profile_id UUID NOT NULL
 );
